Add coverage xml parser. Add coverage module that choose between the TXT or XML report based on the file extension.
