+++ .github/workflows/coverage-xml-test.yml
@@ -1,0 +1,22 @@
+name: Test Coverage Xml
+on:
+  pull_request:
+jobs:
+  coverage-xml-test:
+    name: Coverage Xml Test
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Pytest coverage comment
+        id: coverageComment
+        uses: MishaKav/pytest-coverage-comment@main
+        with:
+          pytest-coverage-path: ./data/coverage.xml
+          junitxml-path: ./data/pytest_1.xml
+
+      - name: Check the output coverage
+        run: |
+          echo "Coverage Report - ${{ steps.coverageComment.outputs.coverage }}"
+          echo "Coverage Color - ${{ steps.coverageComment.outputs.color }}"
+          echo "Coverage Html - ${{ steps.coverageComment.outputs.coverageHtml }}"
