+++ src/cli.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const { getCoverageReport } = require('./parse');
+const { getCoverageReport } = require('./coverage');
 const {
   getSummaryReport,
   getParsedXml,
@@ -8,11 +8,11 @@ const {
 } = require('./junitXml');
 const { getMultipleReport } = require('./multiFiles');
 
-/*  
+/*
   Usefull git commands
-  git tag -a -m "Export coverage example" v1.1.7 && git push --follow-tags 
-  git tag -d v1.0 
-  git tag -d origin v1.0  
+  git tag -a -m "Export coverage example" v1.1.7 && git push --follow-tags
+  git tag -d v1.0
+  git tag -d origin v1.0
 
   # remove all workflows from repo
   gh api repos/MishaKav/pytest-coverage-comment/actions/runs \
